float cost to double

+++ src/structures/operations/hash-join.h
@@ -7,9 +7,9 @@
 class HashJoin : public Operation{
     public:
         HashJoin(bool foreign_key);
-        virtual float getStartCost(Database* database);
-        virtual float getRuntimeCost(Database* database);
-        virtual float getNt();
+        virtual double getStartCost(Database* database);
+        virtual double getRuntimeCost(Database* database);
+        virtual double getNt();
     private:
         int nt;
 };
