+++ src/structures/operations/or-union.h
@@ -7,9 +7,9 @@
 class OrUnion : public Operation{
     public:
         OrUnion();
-        virtual float getStartCost(Database* database);
-        virtual float getRuntimeCost(Database* database);
-        virtual float getNt();
+        virtual double getStartCost(Database* database);
+        virtual double getRuntimeCost(Database* database);
+        virtual double getNt();
     private:
     
 };
